refactor(frontend): split up authentication logic
