+++ frontend/src/components/AppLayout.tsx
@@ -1,68 +1,50 @@
 "use client";
 
 import { ApolloProvider } from "@apollo/client";
-import type { EventMessage, AuthenticationResult } from "@azure/msal-browser";
-import { InteractionType , PublicClientApplication, EventType } from "@azure/msal-browser";
-import { MsalProvider, MsalAuthenticationTemplate } from "@azure/msal-react";
+import { InteractionType } from "@azure/msal-browser";
+import { useMsal, useMsalAuthentication } from "@azure/msal-react";
 import { App, Layout } from "antd";
 import { useEffect, useState } from "react";
 
 import { AppNavbar } from "./AppNavbar";
 import { createApolloClient } from "@/lib/apollo";
-import { loginRequest, msalConfig } from "@/lib/auth";
+import { loginRequest } from "@/lib/auth";
 import type { LayoutProps } from "@/utils/layout";
 
 
 
-const pca = new PublicClientApplication(msalConfig);
-
-
 
 export function AppLayout({ children }: LayoutProps) {
-
-    const accounts = pca.getAllAccounts();
-    if (accounts.length > 0) {
-        pca.setActiveAccount(accounts[0]);
-    } else {
-        console.warn("No Azure accounts logged in");
+    const msal = useMsal();
+    const account = msal.accounts[0];
+    if(!account) {
+        throw new Error("No azure accounts logged in!");
     }
 
-    pca.addEventCallback((event: EventMessage) => {
-        if (event.eventType === EventType.LOGIN_SUCCESS && event.payload) {
-            const payload = event.payload as AuthenticationResult;
-            pca.setActiveAccount(payload.account);
-        }
-    });
+    const auth = useMsalAuthentication(InteractionType.Redirect, loginRequest, account);
+    const [token, setToken] = useState<string | undefined>(auth.result?.idToken);
 
-    // TODO: consider using `pca.setNavigationClient()` for Next-based routing from MSAL.
-
-    const [token, setToken] = useState<string | undefined>();
     useEffect(() => {
-        pca.acquireTokenSilent(loginRequest).then((res) => setToken(res.accessToken));
-    });
+        if(token) { return; }
+        auth.acquireToken(InteractionType.Silent, loginRequest).then((res) => setToken(res?.idToken));
+    }, [auth, token]);
 
     if(!token) { return <></>; }
+
     const apolloClient = createApolloClient(token);
 
     return (
-        <MsalProvider instance={pca}>
-            <ApolloProvider client={apolloClient}>
-                <App>
-                    <MsalAuthenticationTemplate
-                        interactionType={InteractionType.Redirect}
-                        authenticationRequest={loginRequest}
-                    >
-                        <Layout className="h-screen">
-                            <Layout.Header>
-                                <AppNavbar/>
-                            </Layout.Header>
-                            <Layout.Content>
-                                {children}
-                            </Layout.Content>
-                        </Layout>
-                    </MsalAuthenticationTemplate>
-                </App>
-            </ApolloProvider>
-        </MsalProvider>
+        <ApolloProvider client={apolloClient}>
+            <App>
+                <Layout className="h-screen">
+                    <Layout.Header>
+                        <AppNavbar/>
+                    </Layout.Header>
+                    <Layout.Content>
+                        {children}
+                    </Layout.Content>
+                </Layout>
+            </App>
+        </ApolloProvider>
     );
 }
